Jk I figured it out
I didn't realize the LinkedListNode was a class within the package. Once I figured that out, I knew how to work it :D

Solved a few problems with help from solutions manual! I still think I don't entirely understand what is being passed in. Is it always the head? The book mentions that might not always be the case.